Refactor: 코드 전체 폴더링 수정
