--- app/src/main/java/com/sopt/now/compose/BottomNavigationItem.kt
@@ -1,8 +1,0 @@
-package com.sopt.now.compose
-
-import androidx.compose.ui.graphics.vector.ImageVector
-
-data class BottomNavigationItem (
-    val icon: ImageVector,
-    val label: String
-)
