--- app/src/main/java/com/sopt/now/compose/ContextExtensions.kt
@@ -1,8 +1,0 @@
-package com.sopt.now.compose
-
-import android.content.Context
-import android.widget.Toast
-
-fun Context.showToast(message: String) {
-    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
-}
