--- app/src/main/java/com/sopt/now/compose/activity/LoginActivity.kt
@@ -1,171 +1,0 @@
-package com.sopt.now.compose.activity
-
-import android.app.Activity
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import android.util.Log
-import android.widget.Toast
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.viewModels
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.lifecycle.Observer
-import com.sopt.now.compose.LoginViewModel
-import com.sopt.now.compose.R
-import com.sopt.now.compose.ServicePool.authService
-import com.sopt.now.compose.home.HomeViewModel
-import com.sopt.now.compose.request.RequestLoginDto
-import com.sopt.now.compose.response.ResponseLoginDto
-import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
-import com.sopt.now.compose.LoginState
-import com.sopt.now.compose.user.UserInfo
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-
-class LoginActivity : ComponentActivity() {
-
-    private val viewModel by viewModels<LoginViewModel>()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            NOWSOPTAndroidTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    LoginPage(viewModel)
-                }
-            }
-        }
-
-        viewModel.loginState.observe(this, Observer { state ->
-            when {
-                state.isSuccess -> {
-                    Toast.makeText(this, "로그인 성공: ${state.message}", Toast.LENGTH_SHORT).show()
-                    startActivity(Intent(this, MainActivity::class.java))
-                    finish()
-                }
-                else -> {
-                    Toast.makeText(this, "로그인 실패: ${state.message}", Toast.LENGTH_SHORT).show()
-                }
-            }
-        })
-    }
-}
-
-@Composable
-fun LoginPage(viewModel: LoginViewModel) {
-    var id by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-
-    val context = LocalContext.current
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(30.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Spacer(modifier = Modifier.height(40.dp))
-
-        Text(
-            text = stringResource(id = R.string.login_title),
-            fontSize = 30.sp,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier.align(Alignment.CenterHorizontally),
-        )
-        Spacer(modifier = Modifier.height(50.dp))
-
-        Text(
-            text = stringResource(id = R.string.id),
-            fontSize = 20.sp,
-            fontWeight = FontWeight.Bold,
-            textAlign = TextAlign.Start,
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(30.dp))
-        TextField(
-            value = id,
-            onValueChange = { id = it },
-            placeholder = { Text(stringResource(id = R.string.id_hint)) },
-            modifier = Modifier.fillMaxWidth(),
-            keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Next),
-        )
-        Spacer(modifier = Modifier.height(40.dp))
-        Text(
-            text = stringResource(id = R.string.pwd),
-            fontSize = 20.sp,
-            fontWeight = FontWeight.Bold,
-            textAlign = TextAlign.Start,
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.height(30.dp))
-        TextField(
-            value = password,
-            onValueChange = { password = it },
-            placeholder = { Text(stringResource(id = R.string.pwd_hint)) },
-            modifier = Modifier.fillMaxWidth(),
-            visualTransformation = PasswordVisualTransformation(),
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
-        )
-        Spacer(modifier = Modifier.height(200.dp))
-
-        Button(
-            onClick = {
-                viewModel.login(context, id, password)
-            },
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text(
-                text = stringResource(id = R.string.btn_login),
-                fontSize = 20.sp,
-                fontWeight = FontWeight.Bold
-            )
-        }
-        Spacer(modifier = Modifier.height(20.dp))
-
-        Button(
-            onClick = {
-                context.startActivity(Intent(context, SignUpActivity::class.java))
-            },
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text(
-                text = stringResource(id = R.string.btn_sign_up),
-                fontSize = 20.sp,
-                fontWeight = FontWeight.Bold
-            )
-        }
-    }
-}
-
