+++ app/src/main/java/com/sopt/now/compose/presentation/login/LoginState.kt
@@ -1,0 +1,6 @@
+package com.sopt.now.compose.presentation.login
+
+data class LoginState(
+    val isSuccess: Boolean,
+    val message: String
+)
