+++ app/src/main/java/com/sopt/now/compose/presentation/user/UserInfo.kt
@@ -1,0 +1,11 @@
+package com.sopt.now.compose.presentation.user
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class UserInfo(
+    val authenticationId: String,
+    val nickname: String,
+    val phone: String
+) : Parcelable
