--- app/src/main/java/com/sopt/now/compose/response/ResponseLoginDto.kt
@@ -1,12 +1,0 @@
-package com.sopt.now.compose.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseLoginDto(
-    @SerialName("code")
-    val code: Int,
-    @SerialName("message")
-    val message: String,
-)
