--- app/src/main/java/com/sopt/now/compose/search/SearchScreen.kt
@@ -1,9 +1,0 @@
-package com.sopt.now.compose.search
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun SearchScreen() {
-
-}
