--- app/src/main/java/com/sopt/now/compose/user/UserInfo.kt
@@ -1,11 +1,0 @@
-package com.sopt.now.compose.user
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class UserInfo(
-    val authenticationId: String,
-    val nickname: String,
-    val phone: String
-) : Parcelable
