+++ app/src/main/java/com/sopt/now/compose/util/BottomNavigationItem.kt
@@ -1,0 +1,8 @@
+package com.sopt.now.compose.util
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class BottomNavigationItem (
+    val icon: ImageVector,
+    val label: String
+)
