+++ app/src/main/java/com/sopt/now/compose/util/ContextExtensions.kt
@@ -1,0 +1,8 @@
+package com.sopt.now.compose.util
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.showToast(message: String) {
+    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+}
